pyproject.toml
[project]
name = "sit-mas-simulation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "keras-tuner>=1.4.7",
    "matplotlib>=3.10.7",
    "notebook>=7.4.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
]
